Extract session viewer subscription
